Remove substitution from RequireJS wrapper template
Rather than substituting the baseUrl, path to require-cs, and the
content of the spec file into the template, treat the template as a
setup script which is run before the spec to setup a new context.

To accomplish this, pass baseUrl and csPath as part of the context and
append rather than substitute the spec file content.

This is intended to make the RequireJS spec running process simpler
before exposing the functionality to users in a subsequent commit.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>