Add test for RequireJS setup file being run
Note:  The tests of the RequireJS setup file have been added to the
spec-requirejs tests (which include the CoffeeScript option).  This
works fine, as the setup file is specifically designed to make the most
minimal changes to the default setup file, bug it may still be
preferable to isolate the setup file tests from the others in the long
run.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>